requirements.txt
Flask==2.0.3
Flask-SQLAlchemy==2.5.1
Werkzeug==2.0.3
PyJWT==2.3.0
